Advancement Contest Datapack/requirements.txt
nbtlib==2.0.4
numpy==2.3.1
Pillow >= 10.0.0
NumPy >= 1.24